test an issue when OVAL and SCE checks are present in a rule
This commit add an Hello world like test that always pass in
file_permissions_unauthorized_sgid rule.

When both OVAL and SCE check are present, the check fails and no output
from OVAL nor SCE are displayed. The SCE script is not included in the
datastream.

Expected result: the check should pass, and both output from OVAL and
SCE check should be displayed.